% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kripp.boot.R
\name{kripp.boot}
\alias{kripp.boot}
\title{Bootstrap Krippendorff's Alpha}
\usage{
kripp.boot(x, iter = 2000, probs = c(.025, .975), 
           method = c("nominal", "ordinal", "interval", "ratio"))
}
\arguments{
\item{x}{is a matrix with rows (in R: observations) corresponding to judges and
columns (in R: variables) corresponding to rated objects. Should be numeric, with NAs for missing data.}

\item{iter}{the number of iterations for bootstrapping.}

\item{probs}{a vector of probabilities for which confidence values are computed.}

\item{method}{the metric used to calculate the difference function.
"nominal", "ordinal", "interval", and "ratio" are currently implemented.}
}
\value{
A list containing the following components:

\item{$mean.alpha}{the mean value of all bootstrapped alpha replicates}

\item{$alpha}{a vector of bootstrapped alphas}

\item{$upper}{upper alpha value for given probabilities}

\item{$lower}{lower alpha value for given probabilities}

\item{$raters}{number of raters used in calculating alpha}

\item{$iter}{number of bootstrap replications}

\item{$probs}{vector of probabilities used}

\item{$size}{number of items used in calculating alpha}
}
\description{
This function implements Prof. Klaus Krippendorff's algorithm for bootstrapping 
the Krippendorff's alpha coefficient. It computes confidence values (reliability 
estimates) for the given probabilities.
}
\examples{
# Krippendorff's "C" data (2011, 2)
nmm<-matrix(c(1,1,NA,1,2,2,3,2,3,3,3,3,3,3,3,3,2,2,2,2,1,2,3,4,4,4,4,4,
1,1,2,1,2,2,2,2,NA,5,5,5,NA,NA,1,1,NA,NA,3,NA),nrow=4)

# assume default nominal classification with 2000 replicates
kripp.boot(nmm)

# nominal classification with 5000 replicates
kripp.boot(nmm, iter=5000)

# ordinal classification with 5000 replicates
kripp.boot(nmm, iter=5000, method="ordinal")
}
\references{
Krippendorff, K. (2011). Computing Krippendorff's Alpha-Reliability. Retrieved from http://repository.upenn.edu/asc_papers/43
}
\seealso{
Algorithm for bootstrapping a distribution of alpha (http://www.afhayes.com/public/alphaboot.pdf)

Andrew F. Hayes's SPSS code (http://www.afhayes.com/public/kalpha.sps)

Krippendorff, K. (2012). Content analysis: An introduction to its methodology. Sage.
}
\author{
Polina Proutskova (proutskova@googlemail.com)

Mike Gruszczynski (mikewgruz@gmail.com)
}
